별점 리뷰 작성 및 평균 별점 user테이블에 저장되는 로직 수정

+++ gomen/src/main/java/com/tuna/gomen/user/command/entity/User.java
@@ -63,8 +63,7 @@ public class User {
     private int violationScore = 0;
 
     @Column(name = "average_rating")
-    private Float averageRating;
-
+    private Double averageRating;
 
 }
